refactor: being organize types
