+++ packages/chevrotain/package.json
@@ -73,6 +73,7 @@
   "dependencies": {
     "@chevrotain/types": "^9.1.0",
     "@chevrotain/gast": "^9.1.0",
+    "@chevrotain/cst-dts-gen": "^9.1.0",
     "@chevrotain/utils": "^9.1.0",
     "regexp-to-ast": "0.5.0",
     "lodash": "4.17.21"
